+++ AlchemistCode/Relics/GlowingShard.cs
@@ -5,7 +5,7 @@ using MegaCrit.Sts2.Core.Models.Powers;
 
 namespace Alchemist.AlchemistCode.Relics;
 
-public class ToxicShard : AlchemistRelic
+public class GlowingShard : AlchemistRelic
 {
     public override RelicRarity Rarity => RelicRarity.Uncommon;